Added constraints to the input
Added constraints to the input, so that the user can only use letters when entering first and last name, also he can only use numbers between 1 and 10 to enter the grades. Lastly I capitalized all the letters of the name and last name after user input is completed for a cleaner look in the output.